requirements.txt
# Core PyTorch packages (installed first)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
beartype==0.20.2
beautifulsoup4==4.13.4
cachetools==5.5.2
colored==2.3.0
cycler==0.12.1
datasets==3.5.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
humanize==4.12.3
inquirerpy==0.3.4
jiter==0.9.0
matplotlib
multidict==6.4.3
multiprocess==0.70.16
numpy
packaging==25.0
pandas
pfzy==0.3.4
pillow>=9.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml>=5.4.1
qwen-vl-utils==0.0.11
regex==2024.11.6
requests>=2.32.3
safetensors==0.5.3
scipy
seaborn==0.13.2
slurm-gpustat==0.0.15
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.21.1
transformers==4.51.0
typing-inspection==0.4.0
typing_extensions>=4.13.2
tzdata==2025.2
urllib3>=1.26.5
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
